Fix fish user config restoration
