+++ cmuxTests/GhosttyConfigTests.swift
@@ -5846,88 +5846,6 @@ final class ZshShellIntegrationHandoffTests: XCTestCase {
         XCTAssertEqual(result.stdout, "\u{1B}[>m\u{1B}[<8u")
     }
 
-    func testFishIntegrationConfigParsesWhenFishIsAvailable() throws {
-        let fishExecutable = try requireFishExecutable()
-        let repoRoot = URL(fileURLWithPath: #filePath)
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-        let integrationPath = repoRoot.appendingPathComponent("Resources/shell-integration/fish/config.fish")
-
-        let result = runFishProcess(
-            executablePath: fishExecutable,
-            arguments: ["-n", integrationPath.path],
-            environment: ProcessInfo.processInfo.environment,
-            timeout: 5
-        )
-
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-    }
-
-    func testFishIntegrationSourcesUserConfigAndReportsShellState() throws {
-        _ = try requireFishExecutable()
-        let fileManager = FileManager.default
-        let root = fileManager.temporaryDirectory
-            .appendingPathComponent("cmux-fish-shell-integration-\(UUID().uuidString)")
-        let userConfigHome = root.appendingPathComponent("user-config", isDirectory: true)
-        let userFishConfig = userConfigHome.appendingPathComponent("fish/config.fish", isDirectory: false)
-        let logPath = root.appendingPathComponent("send.log", isDirectory: false)
-
-        try fileManager.createDirectory(
-            at: userFishConfig.deletingLastPathComponent(),
-            withIntermediateDirectories: true
-        )
-        defer { try? fileManager.removeItem(at: root) }
-
-        try """
-        set -gx CMUX_USER_FISH_CONFIG_SOURCED 1
-        """.write(to: userFishConfig, atomically: true, encoding: .utf8)
-
-        let result = try runInteractiveFish(
-            command: """
-            functions -e _cmux_socket_is_unix
-            function _cmux_socket_is_unix
-                return 0
-            end
-            functions -e _cmux_send_bg
-            function _cmux_send_bg
-                printf '%s\\n' $argv[1] >> "$CMUX_TEST_LOG"
-            end
-            set -g _CMUX_TTY_NAME ttys777
-            set -g _CMUX_TTY_REPORTED 0
-            _cmux_preexec
-            _cmux_prompt
-            printf 'USER_CONFIG=%s\\n' "$CMUX_USER_FISH_CONFIG_SOURCED"
-            cat "$CMUX_TEST_LOG"
-            """,
-            extraEnvironment: [
-                "CMUX_FISH_CONFIG_HOME": userConfigHome.path,
-                "CMUX_TEST_LOG": logPath.path,
-                "CMUX_SOCKET_PATH": root.appendingPathComponent("cmux-test.sock").path,
-                "CMUX_TAB_ID": "11111111-1111-1111-1111-111111111111",
-                "CMUX_PANEL_ID": "22222222-2222-2222-2222-222222222222",
-            ]
-        )
-
-        XCTAssertTrue(result.stdout.contains("USER_CONFIG=1"), result.stdout)
-        XCTAssertTrue(
-            result.stdout.contains("report_tty ttys777 --tab=11111111-1111-1111-1111-111111111111 --panel=22222222-2222-2222-2222-222222222222"),
-            result.stdout
-        )
-        XCTAssertTrue(
-            result.stdout.contains("report_shell_state running --tab=11111111-1111-1111-1111-111111111111 --panel=22222222-2222-2222-2222-222222222222"),
-            result.stdout
-        )
-        XCTAssertTrue(
-            result.stdout.contains("ports_kick --tab=11111111-1111-1111-1111-111111111111 --panel=22222222-2222-2222-2222-222222222222 --reason=command"),
-            result.stdout
-        )
-        XCTAssertTrue(
-            result.stdout.contains("report_shell_state prompt --tab=11111111-1111-1111-1111-111111111111 --panel=22222222-2222-2222-2222-222222222222"),
-            result.stdout
-        )
-    }
-
     private func runInteractiveZsh(cmuxLoadGhosttyIntegration: Bool) throws -> String {
         try runInteractiveZsh(
             cmuxLoadGhosttyIntegration: cmuxLoadGhosttyIntegration,
@@ -6273,120 +6191,6 @@ final class ZshShellIntegrationHandoffTests: XCTestCase {
         return (
             stdout: output.trimmingCharacters(in: .whitespacesAndNewlines),
             stderr: error.trimmingCharacters(in: .whitespacesAndNewlines)
-        )
-    }
-
-    private struct FishProcessRunResult {
-        let status: Int32
-        let stdout: String
-        let stderr: String
-        let timedOut: Bool
-    }
-
-    private func requireFishExecutable() throws -> String {
-        let candidates = [
-            "/opt/homebrew/bin/fish",
-            "/usr/local/bin/fish",
-            "/usr/bin/fish",
-            "/bin/fish",
-        ]
-        guard let fish = candidates.first(where: { FileManager.default.isExecutableFile(atPath: $0) }) else {
-            throw XCTSkip("fish is not installed")
-        }
-        return fish
-    }
-
-    private func runInteractiveFish(
-        command: String,
-        extraEnvironment: [String: String] = [:]
-    ) throws -> (stdout: String, stderr: String) {
-        let fishExecutable = try requireFishExecutable()
-        let fileManager = FileManager.default
-        let root = fileManager.temporaryDirectory
-            .appendingPathComponent("cmux-fish-shell-integration-\(UUID().uuidString)")
-        try fileManager.createDirectory(at: root, withIntermediateDirectories: true)
-        defer { try? fileManager.removeItem(at: root) }
-
-        let repoRoot = URL(fileURLWithPath: #filePath)
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-        let integrationDir = repoRoot.appendingPathComponent("Resources/shell-integration", isDirectory: true)
-        let testIntegrationDir = root.appendingPathComponent("shell-integration", isDirectory: true)
-        try fileManager.copyItem(at: integrationDir, to: testIntegrationDir)
-
-        var environment: [String: String] = [
-            "HOME": root.path,
-            "TERM": "xterm-256color",
-            "SHELL": fishExecutable,
-            "USER": NSUserName(),
-            "XDG_CONFIG_HOME": testIntegrationDir.path,
-            "CMUX_SHELL_INTEGRATION": "1",
-            "CMUX_SHELL_INTEGRATION_DIR": testIntegrationDir.path,
-        ]
-        for (key, value) in extraEnvironment {
-            environment[key] = value
-        }
-
-        let result = runFishProcess(
-            executablePath: fishExecutable,
-            arguments: ["-i", "-c", command],
-            environment: environment,
-            timeout: 5
-        )
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-        return (
-            stdout: result.stdout.trimmingCharacters(in: .whitespacesAndNewlines),
-            stderr: result.stderr.trimmingCharacters(in: .whitespacesAndNewlines)
-        )
-    }
-
-    private func runFishProcess(
-        executablePath: String,
-        arguments: [String],
-        environment: [String: String],
-        timeout: TimeInterval
-    ) -> FishProcessRunResult {
-        let process = Process()
-        let stdoutPipe = Pipe()
-        let stderrPipe = Pipe()
-        process.executableURL = URL(fileURLWithPath: executablePath)
-        process.arguments = arguments
-        process.environment = environment
-        process.standardInput = FileHandle.nullDevice
-        process.standardOutput = stdoutPipe
-        process.standardError = stderrPipe
-
-        do {
-            try process.run()
-        } catch {
-            return FishProcessRunResult(
-                status: -1,
-                stdout: "",
-                stderr: String(describing: error),
-                timedOut: false
-            )
-        }
-
-        let exitSignal = DispatchSemaphore(value: 0)
-        DispatchQueue.global(qos: .userInitiated).async {
-            process.waitUntilExit()
-            exitSignal.signal()
-        }
-
-        let timedOut = exitSignal.wait(timeout: .now() + timeout) == .timedOut
-        if timedOut {
-            process.terminate()
-            _ = exitSignal.wait(timeout: .now() + 1)
-        }
-
-        let stdout = String(data: stdoutPipe.fileHandleForReading.readDataToEndOfFile(), encoding: .utf8) ?? ""
-        let stderr = String(data: stderrPipe.fileHandleForReading.readDataToEndOfFile(), encoding: .utf8) ?? ""
-        return FishProcessRunResult(
-            status: process.terminationStatus,
-            stdout: stdout,
-            stderr: stderr,
-            timedOut: timedOut
         )
     }
