+++ cmuxTests/GhosttyTerminalStartupEnvironmentTests.swift
@@ -138,10 +138,15 @@ struct GhosttyTerminalStartupEnvironmentTests {
             ambientEnvironment: [:]
         )
 
-        expectEqual(environment["XDG_CONFIG_HOME"], "/Applications/cmux.app/Contents/Resources/shell-integration")
-        expectEqual(environment["CMUX_FISH_CONFIG_HOME"], "/Users/example/.config")
-        expectTrue(protectedKeys.contains("XDG_CONFIG_HOME"))
-        expectTrue(protectedKeys.contains("CMUX_FISH_CONFIG_HOME"))
+        expectEqual(environment["XDG_CONFIG_HOME"], "/Users/example/.config")
+        expectEqual(
+            environment["CMUX_FISH_INTEGRATION_FILE"],
+            "/Applications/cmux.app/Contents/Resources/shell-integration/fish/config.fish"
+        )
+        expectEqual(environment["CMUX_FISH_USER_CONFIG_ALREADY_LOADED"], "1")
+        expectFalse(protectedKeys.contains("XDG_CONFIG_HOME"))
+        expectTrue(protectedKeys.contains("CMUX_FISH_INTEGRATION_FILE"))
+        expectTrue(protectedKeys.contains("CMUX_FISH_USER_CONFIG_ALREADY_LOADED"))
     }
 
     @Test
@@ -158,10 +163,26 @@ struct GhosttyTerminalStartupEnvironmentTests {
             ambientEnvironment: [:]
         )
 
-        expectEqual(environment["XDG_CONFIG_HOME"], "/Applications/cmux.app/Contents/Resources/shell-integration")
-        expectEqual(environment["CMUX_FISH_CONFIG_HOME"], nil)
-        expectTrue(protectedKeys.contains("XDG_CONFIG_HOME"))
-        expectFalse(protectedKeys.contains("CMUX_FISH_CONFIG_HOME"))
+        expectEqual(environment["XDG_CONFIG_HOME"], "/Applications/cmux.app/Contents/Resources/shell-integration/")
+        expectEqual(
+            environment["CMUX_FISH_INTEGRATION_FILE"],
+            "/Applications/cmux.app/Contents/Resources/shell-integration/fish/config.fish"
+        )
+        expectEqual(environment["CMUX_FISH_USER_CONFIG_ALREADY_LOADED"], "1")
+        expectFalse(protectedKeys.contains("XDG_CONFIG_HOME"))
+        expectTrue(protectedKeys.contains("CMUX_FISH_INTEGRATION_FILE"))
+        expectTrue(protectedKeys.contains("CMUX_FISH_USER_CONFIG_ALREADY_LOADED"))
+    }
+
+    @Test
+    func testManagedFishShellCommandLetsGhosttyApplyLoginExecWrapper() {
+        let command = TerminalSurface.managedFishShellCommand(shell: "/Applications/cmux DEV fishsh.app/fish")
+
+        expectEqual(
+            command,
+            "'/Applications/cmux DEV fishsh.app/fish' -il --init-command 'source \"$CMUX_FISH_INTEGRATION_FILE\"'"
+        )
+        expectFalse(command.hasPrefix("exec "))
     }
 
     @Test
